Convert sidebar toggle to button
