+++ src/globalStyles.js
@@ -11,6 +11,8 @@ export const lightGrey = "#BBB";
 export const extraLightGrey = "#F1F1F1";
 export const brandColor = "#5097BA"; // #5DA8A3 (green) or #5097BA (blue)
 export const sidebarColor = "#F2F2F2"; // #F4F4F4
+export const goColor = "#89B77F"; // green
+export const pauseColor = "#E39B39"; // orange
 
 // http://stackoverflow.com/questions/1895476/how-to-style-a-select-dropdown-with-css-only-without-javascript
 export const sidebarField = {
